LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-initial-target/scroll-initial-target-with-anchor-navigation.tentative-expected.txt



FAIL scroll-initial-target does not override anchor navigation. assert_equals: scroll-initial-target sets initial scroll offset expected 90 but got 0
